Added exceptions for all services

+++ src/main/java/com/codecool/services/AccommodationService.java
@@ -6,6 +6,7 @@ import com.codecool.repositories.AccommodationRepository;
 import com.codecool.repositories.RoomRepository;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -19,31 +20,32 @@ public class AccommodationService {
         this.roomRepository = roomRepository;
     }
 
-    public List<Accommodation> getAllAccommodations(){
+    public List<Accommodation> getAllAccommodations() {
         return accommodationRepository.findAll();
     }
 
-    public List<Accommodation> getAllAccommodationsByCity(String cityName){
-        List<Accommodation> allAccommodations =  accommodationRepository.findAll();
+    public List<Accommodation> getAllAccommodationsByCity(String cityName) {
+        List<Accommodation> allAccommodations = accommodationRepository.findAll();
         return allAccommodations.stream().filter(a -> a.getCity().getName().equals(cityName)).collect(Collectors.toList());
     }
 
-    public void addAccommodation(Accommodation accommodation){
-        if(accommodationRepository.findAll().stream().noneMatch(c -> c.getName().equals(accommodation.getName()))){
-            if(accommodation.getCapacity() >= accommodation.getRooms().size()){
+    public void addAccommodation(Accommodation accommodation) {
+        if (accommodationRepository.findAll().stream().noneMatch(c -> c.getName().equals(accommodation.getName()) && c.getCity().getId().equals(accommodation.getCity().getId()))) {
+            if (accommodation.getCapacity() >= accommodation.getRooms().size()) {
                 accommodationRepository.save(accommodation);
-                try{
-                    for(Room room : accommodation.getRooms()){
+                try {
+                    for (Room room : accommodation.getRooms()) {
                         room.setAccommodation(accommodation);
                         roomRepository.save(room);
                     }
-                }
-                catch (Exception e) {
+                } catch (Exception e) {
                     accommodationRepository.deleteById(accommodation.getId());
                 }
-
+            } else {
+                throw new EntityNotFoundException("Accommodation has more rooms than its capacity");
             }
+        } else {
+            throw new EntityNotFoundException("The hotel already exists in this city");
         }
-
     }
 }
